added more details

+++ src/main/java/com/novavrbe/vrbe/models/charactermodels/CharacterStatistic.java
@@ -3,9 +3,11 @@ package com.novavrbe.vrbe.models.charactermodels;
 import com.novavrbe.vrbe.models.enumerations.Stat;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CharacterStatistic {
     private Stat statName;
     private Integer baseStat;
-    private Integer modified;
+    private BigDecimal modified;
 }
